added background color

+++ js/scripts.js
@@ -3,9 +3,8 @@ var triangle= new Array(3);
 
 var dimmensions=0;
 
-for (dimmensions=0; dimmensions<3; dimmensions++)
-{
-triangle[dimmensions]=(parseInt(prompt('Enter the dimmensions in centimeters!',' ')));
+for (dimmensions=0; dimmensions<3; dimmensions++){
+triangle[dimmensions]=(parseFloat(prompt('Enter the dimmensions in centimeters!',' ')));
 }
 if(triangle[0] === triangle[1] && triangle[0] === triangle[2] && triangle[1] === triangle[2]){
 alert("This triangle is Equilateral");
